Added source folder + Razy.py file
